[Generator] Refactor more string extensions.
Move all the string methods that can be an extension to a static class
(re-use the present one) and add tests.
