+++ src/bgen/Generator.cs
@@ -5924,7 +5924,7 @@ public partial class Generator : IMemberGatherer {
 		if (p == null)
 			return;
 
-		print ($"[Advice ({Quote (p.Message)})]");
+		print ($"[Advice ({p.Message.Quote ()})]");
 	}
 
 	public void PrintRequiresSuperAttribute (ICustomAttributeProvider mi)
@@ -5942,7 +5942,7 @@ public partial class Generator : IMemberGatherer {
 		if (p == null)
 			return;
 
-		print ($"[NotImplemented ({Quote (p.Message)})]");
+		print ($"[NotImplemented ({p.Message.Quote ()})]");
 	}
 
 	public void PrintBindAsAttribute (ICustomAttributeProvider mi, StringBuilder sb = null)
@@ -6958,7 +6958,7 @@ public partial class Generator : IMemberGatherer {
 
 						var sender = pars.Length == 0 ? "this" : pars [0].Name;
 
-						var miname = PascalCase (mi.Name);
+						var miname = mi.Name.PascalCase ();
 						if (miname == previous_miname) {
 							// overloads, add a numbered suffix (it's internal)
 							previous_miname = miname;
@@ -6998,7 +6998,7 @@ public partial class Generator : IMemberGatherer {
 							} else
 								eaname = "<NOTREACHED>";
 
-							print ("var handler = {0};", PascalCase (miname));
+							print ("var handler = {0};", miname.PascalCase ());
 							print ("if (handler != null){");
 							indent++;
 							string eventArgs;
@@ -7035,7 +7035,7 @@ public partial class Generator : IMemberGatherer {
 							if (debug)
 								print ("Console.WriteLine (\"Method {0}.{1} invoked\");", dtype.Name, mi.Name);
 
-							print ("var handler = {0};", PascalCase (miname));
+							print ("var handler = {0};", miname.PascalCase ());
 							print ("if (handler != null)");
 							print ("	return handler ({0}{1});",
 								   sender,
@@ -7098,7 +7098,7 @@ public partial class Generator : IMemberGatherer {
 								print ("if (selHandle.Equals (sel{0}Handle))", mi.Name);
 							}
 							++indent;
-							print ("return {0} != null;", PascalCase (mi.Name));
+							print ("return {0} != null;", mi.Name.PascalCase ());
 							--indent;
 						}
 						print ("return global::" + ns.Messaging + ".bool_objc_msgSendSuper_IntPtr (SuperHandle, " + selRespondsToSelector + ", selHandle);");
@@ -7132,7 +7132,7 @@ public partial class Generator : IMemberGatherer {
 
 						string ensureArg = bta.KeepRefUntil == null ? "" : "this";
 
-						var miname = PascalCase (mi.Name);
+						var miname = mi.Name.PascalCase ();
 						if (miname == prev_miname) {
 							// overloads, add a numbered suffix (it's internal)
 							prev_miname = miname;
@@ -7144,14 +7144,14 @@ public partial class Generator : IMemberGatherer {
 							PrintObsoleteAttributes (mi);
 
 							if (bta.Singleton && mi.GetParameters ().Length == 0 || mi.GetParameters ().Length == 1)
-								print ("public event EventHandler {0} {{", CamelCase (GetEventName (mi)));
+								print ("public event EventHandler {0} {{", GetEventName (mi).CamelCase ());
 							else
-								print ("public event EventHandler<{0}> {1} {{", GetEventArgName (mi), CamelCase (GetEventName (mi)));
+								print ("public event EventHandler<{0}> {1} {{", GetEventArgName (mi), GetEventName (mi).CamelCase ());
 							print ("\tadd {{ Ensure{0} ({1})!.{2} += value; }}", dtype.Name, ensureArg, miname);
 							print ("\tremove {{ Ensure{0} ({1})!.{2} -= value; }}", dtype.Name, ensureArg, miname);
 							print ("}\n");
 						} else {
-							print ("public {0}? {1} {{", GetDelegateName (mi), CamelCase (GetDelegateApiName (mi)));
+							print ("public {0}? {1} {{", GetDelegateName (mi), GetDelegateApiName (mi).CamelCase ());
 							print ("\tget {{ return Ensure{0} ({1})!.{2}; }}", dtype.Name, ensureArg, miname);
 							print ("\tset {{ Ensure{0} ({1})!.{2} = value; }}", dtype.Name, ensureArg, miname);
 							print ("}\n");
@@ -7378,7 +7378,7 @@ public partial class Generator : IMemberGatherer {
 					var safe_name = pi.Name.GetSafeParamName ();
 					print ("public {0} {1} {{ get; set; }}",
 						FormatType (type, pi.ParameterType),
-						Capitalize (safe_name));
+						safe_name.Capitalize ());
 
 					if (comma)
 						ctor.Append (", ");
@@ -7391,7 +7391,7 @@ public partial class Generator : IMemberGatherer {
 				print ("\npublic {0} ({1}) {{", async_type.Item1, ctor); indent++;
 				foreach (var pi in async_type.Item2) {
 					var safe_name = pi.Name.GetSafeParamName ();
-					print ("this.{0} = {1};", Capitalize (safe_name), safe_name);
+					print ("this.{0} = {1};", safe_name.Capitalize (), safe_name);
 				}
 				print ("Initialize ();");
 				indent--; print ("}");
@@ -7558,15 +7558,7 @@ public partial class Generator : IMemberGatherer {
 		}
 		return null;
 	}
-
-	static string Capitalize (string str)
-	{
-		if (str.StartsWith ("@", StringComparison.Ordinal))
-			return char.ToUpper (str [1]) + str.Substring (2);
-
-		return char.ToUpper (str [0]) + str.Substring (1);
-	}
-
+	
 	string GetNotificationCenter (PropertyInfo pi)
 	{
 		var a = AttributeManager.GetCustomAttributes<NotificationAttribute> (pi);
@@ -7621,10 +7613,10 @@ public partial class Generator : IMemberGatherer {
 	{
 		var attrs = AttributeManager.GetCustomAttributes<EventNameAttribute> (pi);
 		if (attrs.Length == 0)
-			return CamelCase (pi.Name).GetSafeParamName ();
+			return pi.Name.CamelCase ().GetSafeParamName ();
 
 		var a = attrs [0];
-		return CamelCase (a.EvtName).GetSafeParamName ();
+		return a.EvtName.CamelCase ().GetSafeParamName ();
 	}
 
 	string RenderArgs (IEnumerable<ParameterInfo> pi)
@@ -7641,17 +7633,7 @@ public partial class Generator : IMemberGatherer {
 	{
 		return parameters.Any (pi => pi.ParameterType.IsByRef);
 	}
-
-	string CamelCase (string ins)
-	{
-		return Char.ToUpper (ins [0]) + ins.Substring (1);
-	}
-
-	string PascalCase (string ins)
-	{
-		return Char.ToLower (ins [0]) + ins.Substring (1);
-	}
-
+	
 	Dictionary<string, bool> skipGeneration = new Dictionary<string, bool> ();
 	string GetEventName (MethodInfo mi)
 	{
@@ -7828,17 +7810,7 @@ public partial class Generator : IMemberGatherer {
 			return t.FullName;
 
 	}
-
-	public static string Quote (string s)
-	{
-		if (s == null)
-			return String.Empty;
-		if (s == string.Empty)
-			return @"""""";
-
-		return $"@\"{s.Replace ("\"", "\"\"")}\"";
-	}
-
+	
 	private static string FormatPropertyInfo (PropertyInfo pi)
 	{
 		return pi.DeclaringType.FullName + " " + pi.Name;
